config(core): Move redis connection default to base config (#679)
Currently Halyard is appending a default for the redis config to
all configs it generates. Let's push this down to the base
config so all users get the same defaults and we don't need to
append this all the time.

Currently the base config uses 'services.redis.connection' as the
default. This doesn't match what is used for other services, nor
what is used for the redis connection in other microservices (which
use services.redis.baseUrl).

This commit changes this to look at 'services.redis.baseUrl' for
consistency. Users who have configured a 'services.redis.connection'
property but not 'services.redis.baseUrl' (and who are depending on
that rather than just setting redis.connection directly in their
gate config) will need to update to account for this change.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>