refactor(config): move Halyard defaults into code and deprecate Halyard-specific config (#735)
* refactor(config): move Halyard defaults into code

Halyard currently defaults the `supportedTypes` list of Datadog, Prometheus, SignalFx, and New Relic accounts to a singleton list including type METRICS_STORE, which is the only reasonable choice for these account types. Let's lift that defaulting into the classes themselves, so that Kleat does not need to set them. While I think this would also be a reasonable default to set for the other metrics-only accounts, this change is limited to those accounts currently configurable using Halyard.

* refactor(config): deprecate Halyard-specific config file

Lifts the redis endpoint from the Halyard-specific config to the base config. This should not impact existing Halyard or non-Halyard users.