fix(provider/cf): Cleaned retry exception statement to reflect sockettimeout more accurately (#4611)
* fix(provider/cf): Cleaned up retry exception statement to reflect sockettimeout more accurately

In the HttpCloudFoundryClient there is retry logic that will retry api calls depending on if the exception thrown falls into the retryable exceptions. In the current logic, if a socket exception is thrown then the retry wont take place. This is intentional as there are likely api calls were retrying on a socket exception could be hazardous. For now, we are just removing the statement that says it will retry as that is misleading. This may be revised down the line to handle specific socket exceptions but for now this will suffice

* fix(provider/cf): Cleaned retry exception statement to reflect sockettimeout more accurately

* cleanup imports

Co-authored-by: Zach Smith <zachsmith@Zachs-MBP-2.attlocal.net>