chore(kubernetes): Bump kubectl version (#4630)
Let's bump to 1.17.x of kubectl for the next Spinnaker release.
Currently Kubernetes is on 1.18.x, but to try to maximize overlap
with the version users are running we'll stay at 1.17.x.

This means that we'll officially support 1.16-1.18 (as kubectl
allows a +/- 1 version skew with the cluster). In practice this
should not cause issues for users on <=1.15 as I don't believe
in practice there are generally changes that break talking to
old clusters. This does allow users on 1.17 to start to use
new fields.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>