refactor(artifacts): Improve null-safety of Artifact class (#667)
* fix(artifacts): Improve null safety of expected artifact

The ExpectedArtifact class is annotated such that id and matchArtifact
are not null, but this is not actually enforced anywhere.

First, explicitly write out the AllArgsConstructor that was generated
by lombok, and move the Builder annotation to it (so the builder uses
that constructor when finally creating the object).

Then update the constructor to have a ParametersAreNullable annotation
as we can't control what gets passed to the constructor, and handle
when we are passed a null id or matchArtifact.

For the id case, we'll just translate to an empty string. For the
matchArtifact case, we'll create an empty matchArtifact. As the
point of a matchArtifact is to contain filters that determine whether
a given artifact matches this expected artifact; by defaulting to no
filters seems like a reasonable behavior here.

* refactor(artifacts): Remove deprecated putMetadata

Last time I updated this class (a few months ago) I added the
Deprecated annotation to this function, which is not used anywhere
in Spinnaker. Removing it now.

* refactor(artifacts): Improve null-safety of Artifact class

The Artifact class contains a number of strings that are often null.
To make the class more null-safe, annotate it with NonnullByDefault and
translate any null values into the empty string in the constructor and
in setters.

In order to do this null -> empty translation, we need to explicitly
write out the constructor and the setters that were generated by
lombok.

We rarely mutate artifacts; when we do it often leads to confusing and
fragile logic because we often pass these many levels down in the call
stack, and also sometimes convert to/from Map<> meaning that it's not
always clear if you're operating on the same artifact or a copy. To
greatly simplify this, I'd like to make this class immutable. Based
on my analysis of where we mutate artifacts, the few places can all
be replaced by either (1) fully specifying the artifact up-front or
(2) returning a copy of the artifact with the desired mutations.

To start this process of making Artifact immutable, I've annotated
all of the setters as Deprecated. Once all of the places where we
call these are updated, I'll come back and actually remove the
setters.

* refactor(artifacts): Remove redundant JsonProperty

These are just explicitly specifying the default values (and in
any case we're using the builder to deserialize).

* refactor(artifacts): Clean up metadata creation

This is a tiny refactor to simplify creating the metadata in
an Artifact builder; rather than check if the Map is null every
time we add a key, just initialize it.

* refactor(artifacts): Make Artifact final

If we want the class to truly be immutable in the future, it will
need to be final. We don't currently extend this anywhere, so
let's make it final.