fix(artifacts): Fix incorrect non-null annotation (#669)
* fix(artifacts): Fix incorrect non-null annotation

Adding NullableByDefault to the Artifact class is also marking the
builder methods as Nullable which is incorrect, and annoying as
the generaly pattern is to chain calls. This is causing all calls
like Artifact.builder().name("abc").build() to have every call in
the chain be highlighted by editors that are aware of these
annotations.

Add a MethodsReturnNonnullByDefault to the builder class, as it will
always return non-null from its methods. (Also remove the
ParametersAreNullableByDefault that is implied by the top-level
annotation on the class.)

* fix(artifacts): Return this from builder putMetadata

This is a minor improvement to the builder API; all the auto-generated
builder functions return this so you can chain them, but I forgot to
do this for putMetadata.