feat(peering): add ability to author custom peering agents (#3701)
* feat(peering): add ability to author custom peering agents

This change adds an extension point (in the form of beans injected into the `PeeringAgent`) to allow peering of custom data in the `orca` [SQL] DB.
Currently, the custom peering agents are called without much context as to what the main peering agent is doing (they are simply called at the end of every agent tick, after all main peering agent logic completes).
This can be changed later if needed but for now I wanted to keep the two as independant and simple as possible.

* fix(custom peering): add init method

* fixup! fix(custom peering): add init method

* fixup! fixup! fix(custom peering): add init method

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>