feat(triggers): Add ability to suppress triggers at runtime (#927)
Adds two dynamic properties:
* `scheduler.suppressTriggers`, and
* `scheduler.compensationJob.suppressTriggers`

This is useful when there are two `echo` clusters (for redundancy) to process events but only one of them should actually trigger.
Processing triggers (e.g. running the schedule) but triggering on only one of the clusters allows both clusters to be "in-sync" and when a switch is needed there will be no (or very few) double triggers.