feat(clouddriver): Add generic cloud operation stage (#3749)
This is the first draft of a generic cloud operation stage. The goal of this
stage, and the surrounding work, is to migrate all cloud provider logic from
orca into Clouddriver. The "vision" of this would be that to add a cloud
provider, or new operations for a cloud provider, would no longer require
modification to Orca; only requiring changes to Clouddriver.

There's still a long way to go on this, but it works enough as-is that I feel
comfortable making a PR.