refactor(stats): add a simpler way to collect more data for stats (#943)
* refactor(stats): convert some of the tests to JUnit/Kotlin

I didn't convert the ones that deal with the stats event content because I'm going to refactor how that works in a later commit.

* feat(stats): add a TelemetryEventDataProvider interface

Implementations can add their own data to the stats Event object. This allows us to pull data from a variety of sources without having to cram all that logic into TelemetryEventListener.

* refactor(stats): convert TelemetryService and TelemetryConfig to Kotlin

Some of the Groovy/Java code gets a little weirder because TelemetryConfigProps is in Kotlin, but after the next commit, all its clients will be in Kotlin, so it'll make more sense.

* refactor(stats): move the data collection out of TelemetryEventListener

Now it is split between two different ExecutionDataProviders. For now, this is a little arbitrary, but I anticipate adding more complicated data providers that do RPCs and things, so this should make those future data providers much simpler to reason about and test.

* fix(stats): fix the default stats logging URL

* fix(stats): don't use the term "whitelist"