feat(stats): Persist the unique id to redis (#947)
* feat(stats): Persist the unique id to redis

Currently the only place that the unique id for a Spinnaker instance
is persisted is in the hal config. This works for users using halyard
as it gets generated the first time it is needed then gets re-used
on subsequent deployments. (It does have the failure mode of copying
the hal config as then the new instance would have the same id as the
old one.)

This doesn't work for users either deploying without Halyard or deploying
via kleat. In this case there is nowhere to store this uniqueid (as
kleat explicitly treats the config only as input).

To solve this problem, we'll store the unique id in redis. On startup,
we'll check to see if there's a value in redis; if there is we'll use
that, if not we'll use the configured value (and persist it).

* fix(stats): Address code review comments

* fix(stats): Remove duplicated code