test(appengine): Add tests demonstrating AppEngine artifact bug (#3772)
* fix(provider/appengine): Temporarily revert fix

This reverts commit 711c21aa346936c6021bf33397cff77614b5450c. After
adding tests to demonstrate the bug and fix, we'll re-apply.

* test(appengine): Add tests demonstrating AppEngine artifact bug

This commit adds tests to demonstrate the regression in the
AppEngine artifact handling. The next commit will fix the bug
and update the tests.

* fix(provider/appengine): Fix for issue spinnaker/spinnaker#5836 (#3755)

groovy.lang.MissingMethodException when referencing github config artifact for GAE deploy. The config artifacts  were of instance of HashMap not an Artifact object. Added logic to translate HashMap to Artifact.

Co-Authored-By: Eric Zimanyi <ezimanyi@google.com>

Co-authored-by: guido9j <james_guido@homedepot.com>