test(artifacts): Add some tests around null metadata values (#702)
I had wanted to make the metadata map an ImmutableMap as it is now
immutable, but this caused some NPEs because it is possible to have
metadata with null values. In order to exercise the current functionality,
add some tests around what happens when you have null values in the
metadata map.

At this point this PR will just add the tests so nobody makes the same
mistake; I may come back later to filter out null values, then make the
map immutable, but for now just adding the tests.