refactor(core): deprecate ProviderVersion (#4721)
* feat(security): deprecate ProviderVersion enum and public methods that use ProviderVersion

* feat(core): deprecate public interfaces and methods that use ProviderVersion, and remove ProviderVersion-dependent logic from private methods

* refactor(alicloud): remove ProviderVersion from AliCloud

Although AliCloudCredentials is a public class, I think it should be safe to remove providerVersion because there are no non-default ProviderVersions configurable with Halyard, and this entire module is not even currently part of the OSS build.

* refactor(google): remove ProviderVersion from Google

* refactor(core): remove acceptsVersion from cloud provider operations

Since we no longer filter by ProviderVersion in core, removes all uses of acceptsVersion(ProviderVersion) except those in public (now deprecated) methods and tests, which will be removed next release.

* refactor(kubernetes): remove ProviderVersion from Kubernetes accounts

* refactor(web): remove providerVersion from CredentialsController

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>