fix(spel): Fix flattening of nested arrays in SpEL evaluation (#720)
* test(spel): Add tests on ExpressionTransform

The next commit will fix a bug in ExpressionTransform; first add some
tests to document the current behavior. In particular, test the
behavior in handling nested maps and nested lists.

* fix(spel): Fix flattening of nested arrays in SpEL evaluation

Currently nested lists are being flattened when evaluating expressions.
This is incorrect, as expression evaluation should not affect the shape
of the data structure; it should only handle evaluating expressions
in the components of that data structure.

There are some situations in the Kubernetes provider where we need
a list of lists in the context; this currently breaks because upon
evaluating expressions these lists get flattened.