fix(test): Fix liveness bug in PooledRequestQueueTest (#4750)
* refactor(test): Convert PooledRequestQueueSpec to Java

The next commit will fix some issues with PooledRequestQueueSpec, but
first convert it to Java as a separate commit so it's easier to see
the changes.

* fix(test): Fix liveness bug in PooledRequestQueueTest

A few times I've encountered the tests hanging when executing these
tests (particularly when my system is under load).

There are a lot of race conditions in these tests, which depend on
small differences between timeouts. Let's fix these to be more
deterministic, and to exit within a reasonable amount of time.

The tests now rely on latches to ensure deterministic execution
order. When we want a task to hold on to the queue until another
task fails, we have the task wait on a latch until the other
task is finished, rather than depend on timeouts.

I wrote a couple of small helper functions at the bottom of the
class which make it easier to work with queue.execute, because
it throws a generic Throwable.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>