feat(core): refactor insight menu from dropdown to buttons (#8424)
The motivation for this change is to better surface an action for the user to
take when they are on the search, project, and application pages. The current
UX obscures the next appropriate action behind a drop down menu that visually
blends into the page, making it difficult for users to know what to do next.

The proposed change is to replace the drop down menu with buttons rendered
appropriately for each page. When on the search page, both create project and
application buttons are rendered, with a call to action for create application.
When on project or application pages, only render those buttons. Finally, if a
cache refresh is needed, render this on all pages when appropriate.

At a technical level, this unifies the three dropdowns which currently depend
on 2 separate angular controllers and a react component. The component has been
used in favor of the angular implementations on both search and project pages.