refactor(kubernetes): exclude manifest from infrastructure interfaces (#8438)
* refactor(kubernetes): exclude manifest from IKubernetesLoadBalancer

As part of ongoing Kubernetes performance improvements, Clouddriver will stop caching the entire Kubernetes resource YAML. In the immediate term, we will stop returning the manifest as part of top-level /loadBalancers, /serverGroups, and /securityGroups endpoints. Let's instead block LB details view loading on a separate /manifests request, and use that to hydrate the template.

* fix(kubernetes): fix alignment of Host Rules and Ingress sections

* refactor(kubernetes): exclude manifest from IKubernetesServerGroup

Also refactors ManifestTrafficService to check for associated services in the `loadBalancers` field of a server group rather than check the server group's manifest's annotations, and removes unnecessary extractEntityTags method which is just re-setting entityTags on the same serverGroup from which it was reading it.

* refactor(kubernetes): exclude manifest from IKubernetesServerGroupManager

* refactor(kubernetes): remove manifest from IKubernetesInstance

* refactor(kubernetes): exclude manifest from IKubernetesSecurityGroup

Also add `account` to interface, because unlike the other infra types, the base type does not include `account` (but we are always getting this back from Clouddriver for Kubernetes SGs).

* fix(kubernetes): include manifest as $interpolation context for annotation-driven UI

The Kubernetes [annotation-driven UI](https://spinnaker.io/guides/user/kubernetes-v2/annotations-ui/) expects the full manifest to be available to hydrate custom UI sections specfied as annotations. Since we already have access to the full manifest in this component off of which we read each annotation, simply include it as part of the context we pass to $interpolate.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>