refactor(kubernetes): remove some unused code (#8458)
* refactor(core): remove unused ManifestService

For some reason, we ended up with a KubernetesManifestService and a similar but unused core ManifestService.

* refactor(kubernetes): remove unused method from KubernetesManifestService

We have now removed all clients of `makeManifestRefresher`, and since we will not be adding additional Angular code, it is unlikely that we would ever need a method that directly mutates some input based on the most recent result of fetching manifests.

* refactor(kubernetes): remove unused interface

There is a duplicate IManifestSubscription interface in the DeployStatus component and no other uses.

* refactor(core): remove unused methods from ManifestWriter

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>