fix(kubernetes): Don't overwrite namespace on instance (#8464)
* fix(kubernetes): Don't overwrite namespace on instance

The namespace is already set on the instance details object;
as we're about to remove region from the instance view (in favor of
zone, which is the correct field for an instance) leaving this code
will cause the namespace to be cleared.

* fix(kubernetes): Use account instead of accountName in details view

Only security groups set accountName; the convention for all others
is to only set account. To prepare for removing accountName from
these two objects, read account instead.