refactor(kubernetes): Remove manifest from view objects (#4769)
* refactor(kubernetes): Push getAccount to implementations

First, getAccountName is only needed for security groups; push
that down only to the security group provider.

Second, getAccount is defined in the base class, but depends on
the abstract method getKey that is defined by implementations.
As this is the only reason we need to have the getKey method,
let's just push getAccount itself down to the implementations
so that we can just access the key directly. In fact, instead
of storing it at all, let's just use the key in the constructor
and discard it as we don't need it for anything else.

Eventually the plan will be to not read this off the key at all
in the constructor as we should already be storing the account
as part of the cache data we're using to construct this.

* refactor(kubernetes): Push fields down from ManifestBasedModel

Most of ManifestBasedModel is lookign up fields on a Manifest;
we want to stop using the manifest for all of these lookups because
in most cases we already store the parsed data as part of the CacheData
item (and in the cases we don't, we should start doing that).

In order to simplify moving to that model, push down all the manifest
logic to the implementing classes. It's a bit more code now, but should
make it much easier to understand. We also remove some methods that aren't
necessary because they were added to the superclass even though they are
only needed by one or two implementations.

Add a KubernetesResource interface that keeps track of fields we're
expecting to find on all kubernetes resources (mapping to the correpsonding
interface in deck). This is just helpful to make sure that we get a
compile-time error if we accidentally remove one of these fields.

* refactor(kubernetes): Pull ImageSummary to a class

This is a fairly large class to be an anonymous class; pull it out
to a real class.

* refactor(kubernetes): Remove manifest from view objects

This commit removes the manifest from the view object models,
instead reading the required fields from the manifest in the
constructor, then discarding the manifest.

This is an intermediate step towards not requiring the manifest
in the constructor at all, and instead using the metadata we store
on the cache object.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>