refactor(kubernetes/pipelines): clean up ManifestStatus and DeployStatus components. (#8466)
* refactor(kubernetes/pipelines): make ManifestStatus props more specific

ManifestStatus component does not require the entire stage, only the account name to pass to one of its children; update the contract accordingly.

* refactor(kubernetes/pipelines): make ManifestStatus functional component

This component does not rely on any React lifecycle hooks, and we are encouraging all new contributions to prefer functional components. The @Overridable annotation only works on class components, and we are not sure if their were ever any custom extensions relying on it, so remove this for now. (Any custom extensions are unlikely to still work as expected anyway after the upcoming refactor in which data fetching is pushed down into each per-Manifest ManifestStatus component, so we can revisit appropriate extension points if anyone lets us know that this broke their custom overrides.)

* refactor(kubernetes/pipelines): remove unnecessary capitalization of Kubernetes kinds

The kind as specified on the manifest already has the correct casing.

* refactor(kubernetes/pipelines): move translation from stage manifest to manifest params out of DeployStatus

Deck currently gets the full manifests it receives from Orca in the execution context, parses some details from those full manifests to create a new request to Clouddriver's /manifests endpoint. Much of this logic was added to disambiguate CRDs of the same kind but different API group. Ideally, Deck would have no knowledge of build-in Kubernetes groups and simply receive all the information it needs to construct a valid request. We'll revisit this later, but for now, let's at least move all the logic into private methods on the KubernetesManifestService rather than having them live on the DeployStatus component. This is also important because in a future commit, we will push the logic that sets up subscriptions against the /manifest endpoint further down the tree so that each ManifestStatus component is responsible for fetching its own data.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>