refactor(gcs): Rewrite the GcsStorageService (#911)
* refactor(gcs): turn the GCS Credentials into a Bean

* refactor(gcs): consolidate the dataFilenames into config constants

* refactor(gcs): move this config-parsing logic into the config object

I doubt we still need this, but I guess it's simple enough to keep just
in case 🤷

* refactor(gcs): get rid of the purge old versions logic

This hasn't been necessary since we switched to updating the metadata
(rather than creating new content) for the last-updated files, back in

* refactor(core): Remove StorageService#ensureBucketExists

No one ever calls this on the base class, only on specific implementations, so there's no reason for it to be a interface method.

* refactor(gcs): Rewrite the GcsStorageService

This rewrite uses the google-cloud-storage, which is a higher-level library than
the google-api-services-storage we were using previously. I believe the old
library is machine-generated from the protos, while the new library is manually
written by the GCP team. In addition to being simpler to use, it also comes with
some best practices baked in.

In particular for us, it includes retry logic. This means we can remove our
hand-written retry logic which both had bugs and was the source of a dependency
on clouddriver.

* chore(gcs): audit/clean up dependencies (and remove clouddriver dep!)

* refactor(gcs): call ensureBucketExists in a @PostConstruct

I think it makes more sense to do this as an
ApplicationListener<ApplicationPreparedEvent>, but it would require some
more refactoring that I'm not up for right now.

* refactor(gcs): address some of the review comments from @ezimanyi

* refactor(gcs): wait .5s to update last-modified before returning

* test(gcs): better thread safety for the multi-threaded test

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>