perf(kubernetes): Always do a live lookup for getManifest (#4778)
* refactor(kubernetes): Inline defaultCacheData

This is a tiny function used in one place; it's clearer to just
have the logic inline rather than pass so many parameters.

Also the calling function never accepts nor returns nulls; annotate
it as such.

* refactor(kubernetes): Clean up null safety in resource relationships

implicitRelationships and ownerReferenceRelationships now no longer
accept any null parameters, and can remove their null-checking logic.

This also finally makes all the parameters to convertAsResource
non-null so we can just annotate with as such.

* refactor(kubernetes): Clean up type-safety/null-safety in job provider

The manifest providers should be ManifestProvider<KubernetesV2Manifest>
so that we don't need to cast the result to a KubernetesV2Manifest.

Also getKubernetesJob should return an Optional as it sometimes returns
null; the calling functions have interface contracts so need to return
null but let's annotate them as nullable.

* perf(kubernetes): Only load metrics if events are requested

For performance reasons we have the includeEvents flag in the
getManifest function because in most cases we don't need events.
We only need events when loading a manfiest from deck so that we
can populate them in the UI.

The same is true for metrics. We don't need metrics when checking
on the status of a deployed manifest, only when loading the UI
tab for the manifest.

Rather than add another flag, let's just consider metrics to be
events and only load them when we've requested events.

* perf(kubernetes): Always do a live lookup for getManifest

Many users managing large Kubernetes deployments with Spinnaker
use the liveManifestCalls flag because it allows deployments to
complete without waiting for a cache refresh cycle. The initial
plan to remove this flag (and revert to using the cache always)
was discussed at the Kubernetes SIG and users did not want to
lose this functionality. In order to simplify the code and reduce
the configuration necessary to get a working, scalable deployment,
let's make this the default given that it has proven itself with
many large users.

This commit replaces the two kubernetes manifest providers with a
single one; it always does a live lookup for getManifest() and
always uses the cache for getClusterAndSortAscending (which was
always the case for users of liveManifestCalls because this was
never implemented in the live provider).

This will not immediately speed up deploys for users using
liveManifestCalls because we will still block on force cache refresh.
We need to do this because there are bugs downstream (ie, with
disabling/deleting older server groups) if the cache is not up to
date. The next step will be to fix this cache refresh strategy and
remove the force cache refresh stage in all cases.

* refactor(kubernetes): Remove fromCacheData

We're always passing includeEvents=false now; when we remove all
the logic that was gated by this flag, it simplifies down to a single
line so let's just inline it.

* refactor(kubernetes): Delete metric caching agent

Now that we only ever do live lookups of metrics in the rare cases
where we need them (when someone has the infrastructure tab open
to a pod) we don't need to metrics caching agent; delete it and
the code supporting it.